refactor passport error into middleware
