+++ server/routes/auth/providers/azure.ts
@@ -16,7 +16,6 @@ import {
   StateStore,
   request,
   getTeamFromContext,
-  PassportError,
 } from "@server/utils/passport";
 
 const router = new Router();
@@ -127,7 +126,7 @@ if (env.AZURE_CLIENT_ID && env.AZURE_CLIENT_SECRET) {
         });
         return done(null, result.user, result);
       } catch (err) {
-        return done(PassportError(err, ctx), null);
+        return done(err, null);
       }
     }
   );
