+++ server/routes/auth/providers/oidc.ts
@@ -18,7 +18,6 @@ import {
   StateStore,
   request,
   getTeamFromContext,
-  PassportError,
 } from "@server/utils/passport";
 
 const router = new Router();
@@ -126,7 +125,7 @@ if (env.OIDC_CLIENT_ID && env.OIDC_CLIENT_SECRET) {
           });
           return done(null, result.user, result);
         } catch (err) {
-          return done(PassportError(err, ctx), null);
+          return done(err, null);
         }
       }
     )
