+++ server/routes/auth/providers/slack.ts
@@ -16,11 +16,7 @@ import {
   Team,
   User,
 } from "@server/models";
-import {
-  getTeamFromContext,
-  PassportError,
-  StateStore,
-} from "@server/utils/passport";
+import { getTeamFromContext, StateStore } from "@server/utils/passport";
 import * as Slack from "@server/utils/slack";
 import { assertPresent, assertUuid } from "@server/validation";
 
@@ -107,7 +103,7 @@ if (env.SLACK_CLIENT_ID && env.SLACK_CLIENT_SECRET) {
         });
         return done(null, result.user, result);
       } catch (err) {
-        return done(PassportError(err, ctx), null);
+        return done(err, null);
       }
     }
   );
